Renamed *em27_diag_autom.py -> *em27_scraper.py
